WIP: Halt support for RISCV

+++ probe-rs/src/debug/mod.rs
@@ -729,7 +729,7 @@ fn get_piece_value(core: &Core, p: &gimli::Piece<DwarfReader>) -> Option<u32> {
         Location::Value { value } => Some(value.to_u64(0xff_ff_ff_ff).unwrap() as u32),
         Location::Register { register } => {
             let val = core
-                .read_core_reg(register.0 as u8)
+                .read_core_reg(register.0 as u16)
                 .expect("Failed to read register from target");
             Some(val)
         }
